requirements.txt
transformers
datasets
evaluate
sacrebleu
tiktoken
sentencepiece
python-dotenv
torch
transformers[torch]